WIP -> Inventory Rework
Itemdatabase -> ajout d'un script editeur pour organiser / ajouter les items

------ Inventory:
Séparations de la logique en plusieures classes distinctes.
-> InventoryRaycastDetector, permet de run la logique du raycast et d'envoyer des évènements
-> Gestion du pool manager avec les SpecificID pour bien stocker les instances d'objets / spécifiques
-> ItemSlot rework, on enlève tous les trucs inutiles et ajout d'une structure qui va s'occuper de stocker les informations nécéssaires
-> PicupHandler, ScrollHandler et DropHandler, séparé dans ces 3 scritps distincts permettant de run les logique des inputs séparément (seul le drophandler est un monobehaviour, les autres on en a pas besoin)

------ En cours / à faire:
-> Avoir le visuel de l'objet actuellement sélectionné dans l'inventaire.
-> Sytème de tri par une fonction Sort
-> Relier les events / delegates et autres fonctionnalitées des interface, par exemple si l'objet est dans l'inventaire ou non, si on tient l'objet dans la main ou non.
-> Relier les objets type "clés" de l'inventaire.
-> Relier les interface IinventoryUsable et IInventoryCallback.
-> Finir l'assignation de l'UI par l'utilisation de l'ItemDB.
-> Commencer à rédiger la Doc puis la v1.

+++ Assets/Gameplay/Items/Databases/ItemDatabase.cs
@@ -5,12 +5,11 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ItemDB_", menuName = "New ItemDB")]
 public class ItemDatabase : ScriptableObject
 {
-    public SerializedDictionary<uint, GameObject> items = new SerializedDictionary<uint, GameObject>();
+    public SerializedDictionary<ushort, GameObject> items = new SerializedDictionary<ushort, GameObject>();
 
-    public IInventoryItem Get(uint ID)
+    public IInventoryItem Get(ushort ID)
     {
         items[ID].TryGetComponent(out IInventoryItem item);
         return item;
     }
-    
 }
